BVCRCSEHT06-main/back/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-JWT-Extended>=4.5.3
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
PyPDF2>=3.0.1
python-docx>=0.8.11
python-multipart>=0.0.6
Werkzeug>=3.0.1
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pillow>=10.1.0
opencv-python>=4.8.0
librosa>=0.10.0
moviepy>=1.0.0
pytesseract>=0.3.10
SpeechRecognition>=3.10.0
pyttsx3>=2.90
requests>=2.31.0
gunicorn>=21.2.0
APScheduler>=3.10.4
cryptography>=41.0.7
PyJWT>=2.8.0
